backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
cryptography==42.0.8
gunicorn==22.0.0
pytest==8.3.2
psycopg[binary]==3.2.13
eth-account==0.13.4
PyNaCl==1.5.0
base58==2.1.1
